pyproject.toml
# pyproject.toml - Modern Python project configuration

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "misc-mcp"
version = "0.2.0"
description = "기타 MCP 레이더 collector and report pipeline"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "AI-Friendly DataHub", email = "zzragida@gmail.com"}
]
keywords = ["radar", "data-collection", "rss", "api", "duckdb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "duckdb>=0.9.0",
    "feedparser>=6.0.0",
    "tenacity>=8.0.0",
    "structlog>=23.0.0",
    "jinja2>=3.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.12.0",
    "python-dateutil>=2.8.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0.0",
    "pybreaker>=1.0.0",
]
korean = [
    "kiwipiepy>=0.17",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML",
    "types-requests",
    "types-python-dateutil",
]

[project.urls]
Homepage = "https://github.com/ai-frendly-datahub/MiscMCPRadar"
Documentation = "https://github.com/ai-frendly-datahub/MiscMCPRadar/tree/main/docs"
Repository = "https://github.com/ai-frendly-datahub/MiscMCPRadar"
Issues = "https://github.com/ai-frendly-datahub/MiscMCPRadar/issues"

[tool.setuptools.packages.find]
include = ["radar", "radar.*", "radar_core", "radar_core.*"]

[tool.setuptools.package-data]
radar = ["templates/**/*"]

# Black formatter configuration
[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.ruff_cache
  | __pycache__
  | venv
  | build
  | dist
)/
'''

# Ruff linter configuration
[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
]
ignore = [
    "E501",  # line too long (handled by black)
]
extend-exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "build",
    "dist",
    "*.egg-info",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-first-party = ["radar"]
force-single-line = false
lines-after-imports = 2

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--verbose",
    "--color=yes",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["."]
relative_files = true
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# MyPy configuration
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
show_error_context = true

[[tool.mypy.overrides]]
module = [
    "feedparser.*",
    "bs4.*",
    "duckdb.*",
]
ignore_missing_imports = true
